# config/platform/generated/researchops_openapi_v1.json
{
  "components": {
    "schemas": {
...
      "CursorPage_GateSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GateSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "CursorPage[GateSummary]",
        "type": "object"
      },
...
      "CursorPage_ReleaseSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReleaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "CursorPage[ReleaseSummary]",
        "type": "object"
      },
...
      "GateEvaluationView": {
        "additionalProperties": false,
        "properties": {
          "adapter_id": {
            "title": "Adapter Id",
            "type": "string"
          },
          "adapter_version": {
            "title": "Adapter Version",
            "type": "integer"
          },
          "blocking": {
            "title": "Blocking",
            "type": "boolean"
          },
          "checks": {
            "items": {},
            "title": "Checks",
            "type": "array"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "evaluation_key": {
            "title": "Evaluation Key",
            "type": "string"
          },
          "evidence_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Artifact Id"
          },
          "evidence_manifest_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Manifest Sha256"
          },
          "expected": {
            "additionalProperties": true,
            "title": "Expected",
            "type": "object"
          },
          "gate_id": {
            "title": "Gate Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "limitations": {
            "items": {},
            "title": "Limitations",
            "type": "array"
          },
          "observed": {
            "additionalProperties": true,
            "title": "Observed",
            "type": "object"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "payload_sha256": {
            "title": "Payload Sha256",
            "type": "string"
          },
          "policy_id": {
            "title": "Policy Id",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Artifact Id"
          },
          "source_contract": {
            "title": "Source Contract",
            "type": "string"
          },
          "source_contracts": {
            "items": {},
            "title": "Source Contracts",
            "type": "array"
          },
          "source_evaluation_ids": {
            "items": {},
            "title": "Source Evaluation Ids",
            "type": "array"
          },
          "source_manifest_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Manifest Sha256"
          }
        },
        "required": [
          "id",
          "evaluation_key",
          "payload_sha256",
          "gate_id",
          "scope_type",
          "scope_id",
          "outcome",
          "blocking",
          "severity",
          "adapter_id",
          "adapter_version",
          "policy_id",
          "policy_version",
          "source_contract",
          "expected",
          "observed",
          "checks",
          "limitations",
          "source_contracts",
          "source_evaluation_ids",
          "evaluated_at"
        ],
        "title": "GateEvaluationView",
        "type": "object"
      },
      "GateSummary": {
        "additionalProperties": false,
        "properties": {
          "blocking": {
            "title": "Blocking",
            "type": "boolean"
          },
          "current_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Evaluation Id"
          },
          "effective_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Status"
          },
          "evaluation_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Outcome"
          },
          "gate_id": {
            "title": "Gate Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "waiver": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiver"
          }
        },
        "required": [
          "id",
          "gate_id",
          "scope_type",
          "scope_id",
          "status",
          "blocking",
          "severity",
          "updated_at"
        ],
        "title": "GateSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReleaseArtifactView": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "release_id": {
            "title": "Release Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "release_id",
          "artifact_id",
          "role",
          "created_at"
        ],
        "title": "ReleaseArtifactView",
        "type": "object"
      },
      "ReleaseDetail": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "limitations": {
            "items": {},
            "title": "Limitations",
            "type": "array"
          },
          "manifest_artifact_id": {
            "title": "Manifest Artifact Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "missing_gates": {
            "items": {
              "type": "string"
            },
            "title": "Missing Gates",
            "type": "array"
          },
          "observed_gates": {
            "items": {
              "type": "string"
            },
            "title": "Observed Gates",
            "type": "array"
          },
          "parent_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Release Id"
          },
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "policy_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Sha256"
          },
          "promoted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted At"
          },
          "release_type": {
            "title": "Release Type",
            "type": "string"
          },
          "required_gates": {
            "items": {
              "type": "string"
            },
            "title": "Required Gates",
            "type": "array"
          },
          "source_commit": {
            "title": "Source Commit",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "superseded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded At"
          }
        },
        "required": [
          "id",
          "release_type",
          "status",
          "manifest_artifact_id",
          "source_commit",
          "limitations",
          "created_at",
          "metadata",
          "required_gates",
          "observed_gates",
          "missing_gates"
        ],
        "title": "ReleaseDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Canonical HTTP control plane for ResearchOps runs, immutable artifacts, quality gates, releases, approvals, MLflow and Prefect.",
    "title": "LLM-XAI ResearchOps Control Plane",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{artifact_id}/releases": {
      "get": {
        "operationId": "list_artifact_releases",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReleaseArtifactView"
                  },
                  "title": "Response List Artifact Releases",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Releases",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/v1/gate-evaluations/{evaluation_id}": {
      "get": {
        "operationId": "get_gate_evaluation",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateEvaluationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluation",
        "tags": [
          "quality-gates"
        ]
      }
    },
    "/api/v1/gates": {
      "get": {
        "operationId": "list_gates",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "gate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gate Id"
            }
          },
          {
            "in": "query",
            "name": "scope_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Type"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Id"
            }
          },
          {
            "in": "query",
            "name": "effective_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Effective Status"
            }
          },
          {
            "in": "query",
            "name": "blocking",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Blocking"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_GateSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Gates",
        "tags": [
          "quality-gates"
        ]
      }
    },
    "/api/v1/gates/{gate_result_id}": {
      "get": {
        "operationId": "get_gate",
        "parameters": [
          {
            "in": "path",
            "name": "gate_result_id",
            "required": true,
            "schema": {
              "title": "Gate Result Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Gate",
        "tags": [
          "quality-gates"
        ]
      }
    },
    "/api/v1/gates/{gate_result_id}/history": {
      "get": {
        "operationId": "get_gate_history",
        "parameters": [
          {
            "in": "path",
            "name": "gate_result_id",
            "required": true,
            "schema": {
              "title": "Gate Result Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GateEvaluationView"
                  },
                  "title": "Response Get Gate History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "History",
        "tags": [
          "quality-gates"
        ]
      }
    },
...
    "/api/v1/releases": {
      "get": {
        "operationId": "list_releases",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "release_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source_commit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Commit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_ReleaseSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Releases",
        "tags": [
          "releases"
        ]
      }
    },
    "/api/v1/releases/{release_id}": {
      "get": {
        "operationId": "get_release",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Release",
        "tags": [
          "releases"
        ]
      }
    },
    "/api/v1/releases/{release_id}/artifacts": {
      "get": {
        "operationId": "list_release_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReleaseArtifactView"
                  },
                  "title": "Response List Release Artifacts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Artifacts",
        "tags": [
          "releases"
        ]
      }
    },
    "/api/v1/releases/{release_id}/gates": {
      "get": {
        "operationId": "list_release_gates",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GateSummary"
                  },
                  "title": "Response List Release Gates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gates",
        "tags": [
          "releases"
        ]
      }
    },
    "/api/v1/releases/{release_id}/promotion-decisions": {
      "get": {
        "operationId": "list_release_promotion_decisions",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromotionDecisionView"
                  },
                  "title": "Response List Release Promotion Decisions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decisions",
        "tags": [
          "releases"
        ]
      }
    },
...
    "/api/v1/scopes/{scope_type}/{scope_id}/gates": {
      "get": {
        "operationId": "list_scope_gates",
        "parameters": [
          {
            "in": "path",
            "name": "scope_type",
            "required": true,
            "schema": {
              "title": "Scope Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GateSummary"
                  },
                  "title": "Response List Scope Gates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scope Gates",
        "tags": [
          "quality-gates"
        ]
      }
    },
...
  }
}